Allow sessions to handle different methods of authentication
The previous session-related classes incorrectly assumed that authentication was
happening via an email / password combination.  This commit allows that
behavior to be configured on a per- `Session` and `SessionController`
basis through the use of `Session.authentication_identifier` and
`SessionsController.persist_authenticated`.  Note that this is a
temporary change to implement this configuration, I hope to do away with
this in future versions.